+++ tst/CTA.Rules.Test/CTATests.cs
@@ -215,13 +215,13 @@ namespace CTA.Rules.Test
             //We don't care about version for CTA-only rules:
             string version = "net5.0";
 
-            var solutionPath = CopySolutionFolderToTemp("MvcMusicStore.sln", downloadLocation);
+            var solutionPath = CopySolutionFolderToTemp("Modernize.Web.sln", downloadLocation);
             var solutionDir = Directory.GetParent(solutionPath).FullName;
 
             FileAssert.Exists(solutionPath);
 
 
-            string projectFile = Directory.EnumerateFiles(solutionDir, "*.csproj", SearchOption.AllDirectories).FirstOrDefault();
+            string projectFile = Directory.EnumerateFiles(solutionDir, "Modernize.Web.Mvc.csproj", SearchOption.AllDirectories).FirstOrDefault();
             FileAssert.Exists(projectFile);
 
             ProjectConfiguration projectConfiguration = new ProjectConfiguration()
@@ -247,24 +247,27 @@ namespace CTA.Rules.Test
             }
 
             var analysisResult = str.ToString();
-            StringAssert.Contains("HtmlEncoder", analysisResult);
+            //StringAssert.Contains("HtmlEncoder", analysisResult);
 
             solutionRewriter.Run(analysisRunResult.ProjectResults.ToDictionary(p => p.ProjectFile, p => p.ProjectActions));
 
             string projectDir = Directory.GetParent(projectFile).FullName;
 
-            var accountControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "AccountController.cs"));
-            var checkoutControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "CheckoutController.cs"));
-            var shoppingCartControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "ShoppingCartController.cs"));
-            var storeManagerControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "StoreManagerController.cs"));
-            var musicStoreEntitiesText = File.ReadAllText(Path.Combine(projectDir, "Models", "MusicStoreEntities.cs"));
-            var shoppingCartText = File.ReadAllText(Path.Combine(projectDir, "Models", "ShoppingCart.cs"));
+            var customerControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "CustomersController.cs"));
+            var productControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "ProductsController.cs"));
 
-            var shoppingCartRemoveViewModel = File.ReadAllText(Path.Combine(projectDir, "ViewModels", "ShoppingCartRemoveViewModel.cs"));
-            var shoppingCartViewModel = File.ReadAllText(Path.Combine(projectDir, "ViewModels", "ShoppingCartViewModel.cs"));
+            //var storeManagerControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "StoreManagerController.cs"));
+            //var accountControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "AccountController.cs"));
+            //var checkoutControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "CheckoutController.cs"));
+            //var shoppingCartControllerText = File.ReadAllText(Path.Combine(projectDir, "Controllers", "ShoppingCartController.cs"));
+            //var musicStoreEntitiesText = File.ReadAllText(Path.Combine(projectDir, "Models", "MusicStoreEntities.cs"));
+            //var shoppingCartText = File.ReadAllText(Path.Combine(projectDir, "Models", "ShoppingCart.cs"));
+
+            //var shoppingCartRemoveViewModel = File.ReadAllText(Path.Combine(projectDir, "ViewModels", "ShoppingCartRemoveViewModel.cs"));
+            //var shoppingCartViewModel = File.ReadAllText(Path.Combine(projectDir, "ViewModels", "ShoppingCartViewModel.cs"));
 
 
-            
+
         }
 
         [Test]
